codecs & packetizers: fix warnings     use size_t instead of int to represent sizes     removes unused parameters from:         block_BytestreamInit()         mpeg4_audio/LOASSyncInfo()         cinepak/cinepak_Getv1()         speex/SendPacket()         subsass/ParseColor()         ffmpeg/postprocess/InitPostproc() & PostprocPict()         dvbsub/YuvaYuvp() & encode_pixel_line_{2,4,8}bp()         faad/DoReordering()
vlc_block_helper.h:
    use size_t, removes unused parameter from block_BytestreamInit()

struct filter_t:
    removes unused parameters from pf_picture_link() & pf_picture_unlink()
    
cmml: fix a segfault (p_item isn't an input item, despite the name) present for some months, proving that nobody did use that code ?

+++ modules/codec/adpcm.c
@@ -68,8 +68,8 @@ struct decoder_sys_t
 {
     enum adpcm_codec_e codec;
 
-    int                 i_block;
-    int                 i_samplesperblock;
+    size_t              i_block;
+    size_t              i_samplesperblock;
 
     audio_date_t        end_date;
 };
@@ -618,7 +618,7 @@ static void DecodeAdpcmDk4( decoder_t *p_dec, int16_t *p_sample,
 {
     decoder_sys_t *p_sys  = p_dec->p_sys;
     adpcm_ima_wav_channel_t channel[2];
-    int                     i_nibbles;
+    size_t                  i_nibbles;
     int                     b_stereo;
 
     b_stereo = p_dec->fmt_in.audio.i_channels == 2 ? 1 : 0;
